+++ siprix_voip_sdk/lib/siprix_voip_sdk.dart
@@ -221,6 +221,17 @@ class CallVideoUpgradedArg {
   }
 }
 
+/// Helper class for handling 'onCallVideoUpgradeRequested' event raised by library
+class CallVideoUpgradeRequestedArg {
+  int callId=0;
+  bool fromMap(Map<dynamic, dynamic> argsMap) {
+    int argsCounter=0;
+    argsMap.forEach((key, value) {
+      if((key == SiprixVoipSdkPlatform.kArgCallId)&&(value is int))     { callId = value; argsCounter+=1; }
+    });
+    return (argsCounter==1);
+  }
+}
 
 /// Helper class for handling 'onCallRedirected' event raised by library
 class CallRedirectedArg {
@@ -388,8 +399,8 @@ class NetStateListener {
 class CallStateListener {
   CallStateListener({this.proceeding, this.incoming, this.incomingPush, this.acceptNotif,
     this.connected, this.terminated, this.dtmfReceived,
-    this.transferred, this.redirected, this.videoUpgraded, this.held, this.switched,
-    this.playerStateChanged});
+    this.transferred, this.redirected, this.videoUpgraded, this.videoUpgradeRequested, 
+    this.held, this.switched, this.playerStateChanged});
 
   ///Triggered by library when changed player state in specific call
   void Function(int playerId, PlayerState s)? playerStateChanged;
@@ -411,6 +422,8 @@ class CallStateListener {
   void Function(int origCallId, int relatedCallId, String referTo)? redirected;
   ///Triggered by library when remote side requested to start send/receive video and request accepted.
   void Function(int callId, bool withVideo)? videoUpgraded;
+  ///Triggered by library when remote side requested to start send/receive video
+  void Function(int callId)? videoUpgradeRequested;
   ///Triggered by library when received DTMF tone from remote side.
   void Function(int callId, int tone)? dtmfReceived;
   ///Triggered by library when local or remote side has put call on hold
@@ -545,6 +558,7 @@ class SiprixVoipSdk {
   static const int kUpgradeToVideoSendRecv = 0;
   static const int kUpgradeToVideoRecvOnly = 1;
   static const int kUpgradeToVideoInactive = 2;
+  static const int kUpgradeToVideoManual   = 3;
 
   ///Error codes constants
   static const int eOK = 0;
@@ -749,6 +763,11 @@ class SiprixVoipSdk {
     return _platform.upgradeToVideo(callId);
   }
 
+  /// Accept requested upgrade of the call's media and specify is allowed to use video
+  Future<void> acceptVideoUpgrade(int callId, bool withVideo) async {
+    return _platform.acceptVideoUpgrade(callId, withVideo);
+  }
+
   /// Stop playing (mute) ringtone.
   Future<void> stopRingtone() async {
     return _platform.stopRingtone();
@@ -900,6 +919,12 @@ class SiprixVoipSdk {
     return _platform.getCallKitCallUUID(sip_callId);
   }
 
+  ///End CallKit call specified by its UUID (Allowed to use only when missed SIP signalling for that call)
+  Future<void>? endCallKitCall(String callkit_CallUUID) {
+    return _platform.endCallKitCall(callkit_CallUUID);
+  }
+
+
   //-//////////////////////////////////////////////////////////////////////////////////////
   //-Android specific implementation
 
@@ -944,6 +969,7 @@ class SiprixVoipSdk {
       case SiprixVoipSdkPlatform.kOnCallTransferred  : _onCallTransferred(argsMap);  break;
       case SiprixVoipSdkPlatform.kOnCallRedirected   : _onCallRedirected(argsMap);   break;
       case SiprixVoipSdkPlatform.kOnCallVideoUpgraded: _onCallVideoUpgraded(argsMap); break;
+      case SiprixVoipSdkPlatform.kOnCallVideoUpgradeRequested: _onCallVideoUpgradeRequested(argsMap); break;
       case SiprixVoipSdkPlatform.kOnCallSwitched     : _onCallSwitched(argsMap);     break;
       case SiprixVoipSdkPlatform.kOnCallHeld         : _onCallHeld(argsMap);         break;
 
@@ -1036,6 +1062,13 @@ class SiprixVoipSdk {
     CallVideoUpgradedArg arg = CallVideoUpgradedArg();
     if(arg.fromMap(argsMap)) {
       callListener?.videoUpgraded?.call(arg.callId, arg.withVideo);
+    }
+  }
+
+  void _onCallVideoUpgradeRequested(Map<dynamic, dynamic> argsMap) {
+    CallVideoUpgradeRequestedArg arg = CallVideoUpgradeRequestedArg();
+    if(arg.fromMap(argsMap)) {
+      callListener?.videoUpgradeRequested?.call(arg.callId);
     }
   }