+++ siprix_voip_sdk/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.31
+- Updated upgradeToVideo implementation. Added new mode 'manual'
+- Added new event 'onCallVideoUpgradeRequested'
+  Triggered when specified 'acc.upgradeToVideo=manual' and remote side requests to add video
+- Added new method 'acceptVideoUpgrade'
+  App has invoke it from 'onCallVideoUpgradeRequested' event handler
+- iOS: Added new method 'endCallKitCall'
+  Example app uses it to end CallKit calls when SIP INVITE hasn't received
+- Android: Added ability to disable displaying activity on lock screen by adding to manifest
+  `<meta-data android:name="com.siprix.DontShowWhenLocked" android:value="true" />`
+  //2025.11.16
+
 ## 1.0.30
 - Added new setting 'acc.upgradeToVideo' and event 'onCallVideoUpgraded'
   //2025.11.07